+++ app/src/main/java/com/leonis/android/adhafera/views/index/PaymentListAdapter.java
@@ -12,14 +12,15 @@ import com.leonis.android.adhafera.models.Payment;
 
 import java.text.SimpleDateFormat;
 import java.util.List;
+import java.util.Locale;
 
 /**
  * Created by leonis on 2018/12/30.
  */
 
-public class PaymentListAdapter extends BaseAdapter {
-    private Context context;
-    private List<Payment> payments;
+class PaymentListAdapter extends BaseAdapter {
+    private final Context context;
+    private final List<Payment> payments;
 
     private class ViewHolder {
         TextView contentView;
@@ -55,9 +56,10 @@ public class PaymentListAdapter extends BaseAdapter {
 
         Payment payment = payments.get(position);
 
-        if (view == null) {
-            LayoutInflater inflater =
-                    (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        LayoutInflater inflater =
+                (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+
+        if (view == null && inflater != null) {
             view = inflater.inflate(R.layout.payment_list_view, parent, false);
 
             TextView contentView = view.findViewById(R.id.payment_content);
@@ -71,20 +73,19 @@ public class PaymentListAdapter extends BaseAdapter {
             holder.categoryView = categoryView;
             holder.priceView = priceView;
             view.setTag(holder);
-
         } else {
             holder = (ViewHolder) view.getTag();
         }
 
         holder.contentView.setText(payment.getContent());
 
-        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd");
+        SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd", Locale.JAPAN);
         holder.dateView.setText(format.format(payment.getDate()));
 
         String[] categories = payment.getCategories();
-        String categoryString = "";
+        StringBuilder categoryString = new StringBuilder();
         for(String category : categories) {
-            categoryString += category + ",";
+            categoryString.append(category).append(",");
         }
         holder.categoryView.setText(categoryString.substring(0, categoryString.length() - 1));